scxtop: fix for building on aarch64+riscv64
Commit 151b1994caad ("scxtop: better naming for kernel worker threads")
introduces some code that does not compile for aarch64/riscv64:

```
error[E0308]: mismatched types
   --> tools/scxtop/src/lib.rs:473:57
    |
473 |                 let next_comm = unsafe { CStr::from_ptr(sched_switch.next_comm.as_ptr()) }
    |                                          -------------- ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected `*const u8`, found `*const i8`
    |                                          |
    |                                          arguments to this function are incorrect
    |
    = note: expected raw pointer `*const u8`
               found raw pointer `*const i8`
```

It seems the bindgen is outputting [i8; 16] for next_comm which means
as_ptr returns a *const i8. Casting this to the platform dependant
c_char (which translates to u8/i8 based on the platform) makes the code
build on all platforms.

Fixes: 151b1994caad ("scxtop: better naming for kernel worker threads")
Signed-off-by: Fredrik Lönnegren <fredrik@frelon.se>